apfsck: fix the width of block and chunk counts
The in-memory spaceman structure is mistakenly using 32-bit types for
chunk and block counts, which would not be enough for very large
containers.  This is not a real issue for now, but get it fixed.

+++ apfsck/spaceman.h
@@ -165,11 +165,11 @@ struct spaceman {
 	u32 sm_chunks_per_cib;
 	u32 sm_cibs_per_cab;
 	u32 sm_cib_count;
-	u32 sm_chunk_count;
+	u64 sm_chunk_count;
 
 	/* Spaceman info measured by the fsck */
-	u32 sm_chunks;	/* Number of chunks */
-	u32 sm_blocks;	/* Number of blocks */
+	u64 sm_chunks;	/* Number of chunks */
+	u64 sm_blocks;	/* Number of blocks */
 	u64 sm_free;	/* Number of free blocks */
 };
